Refactor into a hierarchy builder
